Absorb solar cell models into `models` submodule
